labs/lab_26_git_automation/requirements.txt
pytest==7.4.3
rich==13.7.0
gitpython==3.1.41
python-dotenv==1.0.0
jinja2==3.1.2
requests==2.31.0
pytest-git==1.0.0 